requirements.txt
amqp==5.2.0
asgiref==3.8.1
astroid==3.0.2
attrs==23.2.0
autopep8==1.6.0
billiard==4.2.0
celery==5.3.6
celery-types==0.22.0
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.4.4
cryptography==42.0.5
defusedxml==0.7.1
distlib==0.3.8
dj-rest-auth==5.1.0
Django==4.2.11
django-allauth==0.61.1
django-appconf==1.0.6
django-avatar==8.0.0
django-celery-results==2.5.1
django-extensions==3.1.3
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.14.5
dnspython==2.6.1
drf-spectacular==0.27.1
exceptiongroup==1.2.0
filelock==3.14.0
flake8==7.0.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.6
maturin==1.4.0
mccabe==0.7.0
mypy==1.7.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.0
pillow==10.2.0
pipenv==2023.12.1
platformdirs==4.2.2
pluggy==1.4.0
prompt-toolkit==3.0.43
psutil==5.9.8
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
pylint==3.0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pyperf==2.6.2
pytest==8.1.1
pytest-cov==5.0.0
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
python3-openid==3.2.0
PyJWT==2.8.0
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.4
tomlkit==0.12.3
tomli==2.0.1
types-jsonschema==4.21.0.20240331
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.20240311
types-requests==2.31.0.20240311
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.26.2
wcwidth==0.2.13
